+++ clang/lib/Sema/SemaDecl.cpp
@@ -14065,17 +14065,7 @@ void Sema::ActOnUninitializedDecl(Decl *RealDecl) {
 
     // C++1z [dcl.dcl]p1 grammar implies that an initializer is mandatory.
     if (isa<DecompositionDecl>(RealDecl)) {
-      Preprocessor &PP = getPreprocessor();
-      SourceManager &SM = Context.getSourceManager();
-      LangOptions LO = Context.getLangOpts();
-
-      // Lexer previously checked for '=' and didn't find it
-      // Highlight the token found in its place in the error message
-      Token Tok;
-      Lexer::getRawToken(PP.getLastCachedTokenLocation(), Tok, SM, LO);
-
-      Diag(Tok.getLocation(), diag::err_decomp_decl_requires_init)
-          << Var << Lexer::getSpelling(Tok, SM, LO);
+      Diag(Var->getLocation(), diag::err_decomp_decl_requires_init) << Var;
       Var->setInvalidDecl();
       return;
     }