+++ clang/test/SemaCXX/cxx1z-decomposition.cpp
@@ -121,7 +121,7 @@ void for_range() {
 }
 
 int error_recovery() {
-  auto [foobar]; // expected-error {{requires an initializer}}
+  auto [foobar]; // expected-error {{requires an initializer}} expected-error {{expected initializer before ';'}}
   return foobar_; // expected-error {{undeclared identifier 'foobar_'}}
 }